fix indexing
round -> float

+++ utils/__init__.py
@@ -5,4 +5,4 @@ from . import julian
 def az_to_octant(azimuth):
     octants = ['N', 'NE', 'E', 'SE', 'S', 'SW', 'W', 'NW']
     idx = azimuth / (2*math.pi/8)
-    return octants[round(idx) % 8]
+    return octants[int(round(idx)) % 8]
